feat: add card component
