+++ src/components/list/List.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { Person } from "../../models/Person";
 import "./List.css";
 import Pagination from "../pagination/Pagination";
+import Card from "../card/Card";
 
 interface Props {
     result: Person[];
@@ -12,25 +13,12 @@ interface Props {
 }
 
 const List: React.FC<Props> = ({ result, currentPage, totalPages, onPageChange, onItemClick }) => {
-
-    const extractIdFromUrl = (url: string): number => {
-        const parts = url.split('/');
-        const id = parts[parts.length - 2];
-        return parseInt(id, 10);
-    };
-
-    const handleItemClick = (id: number) => {
-        onItemClick(id);
-    };
-
     return (
         result.length > 0 ?
             <div className="result-list">
                 <ul>
                     {result.map((person) => (
-                        <li key={extractIdFromUrl(person.url)} className='person' onClick={() => handleItemClick(extractIdFromUrl(person.url))}>
-                            <h2>{person.name}</h2>
-                        </li>
+                        <Card key={person.url} data={person} onClick={onItemClick}></Card>
                     ))}
                 </ul>
                 <Pagination currentPage={currentPage} totalPages={totalPages} onPageChange={onPageChange} />
